% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovAncov.R
\name{aovAncov}
\alias{aovAncov}
\title{ANOVA and ANCOVA for analysis of the influence of sex and SVL}
\usage{
aovAncov(data, variaveis, variavel_sexo, variavel_svl)
}
\arguments{
\item{data}{Data frame containing the variables}

\item{variaveis}{Vector with names of dependent variables}

\item{variavel_sexo}{Name of the sex variable}

\item{variavel_svl}{Name of the SVL variable}
}
\value{
A list with the results of the interaction models, sum models, and selected model
}
\description{
Compares models with and without interaction and selects the best based on AIC
}
