% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calculate basic statistics}
\usage{
calc_stats(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A vector with mean, median, standard deviation, and coefficient of variation
}
\description{
Calculate basic statistics
}
