% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkup.R
\name{Chekup}
\alias{Chekup}
\title{Chekup: Normality Analysis with Histograms}
\usage{
Chekup(data, variaveis, pasta_raw, pasta_log)
}
\arguments{
\item{data}{A data frame containing the variables to analyze.}

\item{variaveis}{A character vector with the names of the variables to analyze.}

\item{pasta_raw}{Directory to save the results for raw data.}

\item{pasta_log}{Directory to save the results for log-transformed data.}
}
\value{
A list with paths to the directories where results were saved.
}
\description{
This function performs normality analysis for specified variables in a dataset,
creates histograms, and saves the results in separate directories for raw and log-transformed data.
}
\examples{
\dontrun{
data <- data.frame(var1 = rnorm(100), var2 = runif(100, 1, 10))
Chekup(data, c("var1", "var2"), "raw_results", "log_results")
}
}
