% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutliers.R
\name{getOutliers}
\alias{getOutliers}
\title{Outlier Detection}
\usage{
getOutliers(data, variaveis, variavel_base)
}
\arguments{
\item{data}{Data frame containing the variables}

\item{variaveis}{Vector with the names of the variables to be analyzed}

\item{variavel_base}{Base independent variable for the analysis}
}
\value{
A list with the outlier results and the modified data frame
}
\description{
Identifies outliers based on the base variable and replaces them with NA
}
