% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers by interpolation}
\usage{
remove_outliers(data, dependent_vars, independent_var, distance_threshold)
}
\arguments{
\item{data}{Data frame with dependent and independent variables}

\item{dependent_vars}{Vector with names of dependent variables}

\item{independent_var}{Name of the independent variable}

\item{distance_threshold}{Distance threshold to identify outliers}
}
\value{
A list with the data frame without outliers and a list of found outliers
}
\description{
Remove outliers by interpolation
}
