% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_stats}
\alias{calc_stats}
\title{Calcula estatísticas básicas}
\usage{
calc_stats(x)
}
\arguments{
\item{x}{Um vetor numérico.}
}
\value{
Um vetor com média, mediana, desvio padrão e coeficiente de variação.
}
\description{
Calcula estatísticas básicas
}
